refactoring: change model Vacancy(add field m2m), add module `vacancy_negotiations` and the code was sorted into modules. Add docstring dockstring
