+++ src/api/v1/vacancies/get_vacancies.py
@@ -1,24 +1,18 @@
 import aiohttp
 
 from asgiref.sync import sync_to_async
-from django.core.exceptions import ObjectDoesNotExist
 
+from .utils import _get_user
 from config.vacancies_settings import (
     EXPERIENCE_NAMES,
     LINK_TO_VACANCIES,
-    LINK_TO_VACANCY_REQUEST,
-    ERRORS_TO_BREAK,
-    ERRORS_TO_CONTINUE,
 )
 from dataclasses import dataclass
 from models.models import Vacancy, Notification
 from tbot import main
 from user.models import User
 
-from .exceptions import AuthError, ToContinueError, ToBreakError
-from typing import NoReturn
-
-__all__ = ("vacancies", "send_negotiations")
+__all__ = ("vacancies",)
 
 
 @dataclass
@@ -38,21 +32,6 @@ class ResponseVacancy:
 class InfoNotification:
     id_notification: int
     count_vacancies: int
-
-
-@dataclass
-class SendOutResumeResponse:
-    status_code: int
-    response_json: dict
-
-
-async def _make_request_to_negotiations(url, headers) -> SendOutResumeResponse:
-    async with aiohttp.ClientSession() as session:
-        async with session.post(url=url, headers=headers) as response:
-            return SendOutResumeResponse(
-                status_code=response.status,
-                response_json=await response.json() if response.status != 201 else None,
-            )
 
 
 async def _make_request(url: str) -> dict:
@@ -128,13 +107,14 @@ async def _save_vacancies(
     response_vacancies: list[ResponseVacancy], user: User
 ) -> None:
     for vacancy in response_vacancies:
-        await Vacancy.objects.aget_or_create(user=user, **vacancy.__dict__)
+        _vacancy = await Vacancy.objects.aget_or_create(**vacancy.__dict__)
+        await sync_to_async(_vacancy[0].user.add)(user)
 
 
 async def _parse_vacancies(response_vacancies: dict, exp: str, user: User) -> None:
     vacancies_data = []
     for vacancy in response_vacancies["items"]:
-        vacancies_data.append(_build_data(vacancy, exp))
+        vacancies_data.append(await sync_to_async(_build_data)(vacancy, exp))
     await _save_vacancies(vacancies_data, user)
 
 
@@ -147,7 +127,9 @@ def _get_link_to_vacancies(
 
 
 def _create_notification(user: User) -> InfoNotification:
-    query = list(Vacancy.objects.filter(is_request=False, user=user)[:200])
+    query = list(
+        Vacancy.objects.filter(user=user, vacanciesinfo__is_request=False)[:200]
+    )
     notification = Notification.objects.create(user=user)
     notification.vacancies.set(query)
     return InfoNotification(id_notification=notification.pk, count_vacancies=len(query))
@@ -167,67 +149,6 @@ async def _get_vacancies(user: User) -> None:
             await _parse_vacancies(response_vacancies, exp, user)
 
 
-def _check_response_negotiations(response: SendOutResumeResponse) -> NoReturn | None:
-    try:
-        if response.status_code != 201 and response.response_json is not None:
-            if response.response_json["errors"][0]["value"] in ERRORS_TO_CONTINUE:
-                raise ToContinueError
-            elif response.response_json["errors"][0]["value"] in ERRORS_TO_BREAK:
-                raise ToBreakError
-            else:
-                raise AuthError
-    except KeyError:
-        return ToContinueError
-
-
-def _get_link_to_negotiations(
-    vacancy_id: int, resume_id: str, covering_letter: str
-) -> str:
-    return LINK_TO_VACANCY_REQUEST.format(
-        vacancy_id=vacancy_id,
-        resume_id=resume_id,
-        message=covering_letter,
-    )
-
-
-async def _get_user(user_id) -> User | None:
-    try:
-        return await sync_to_async(User.objects.get)(pk=user_id)
-    except ObjectDoesNotExist:
-        return None
-
-
-def _get_headers(user_access_token) -> dict:
-    return {"Authorization": f"Bearer {user_access_token}"}
-
-
-def _get_notification(id_notification) -> list[Vacancy]:
-    return list(Notification.objects.get(pk=id_notification).vacancies.all())
-
-
-async def _make_negotiations(
-    vacancies_by_notifications: list[Vacancy], user: User, headers: dict
-) -> int:
-    count_valid_requests = 0
-    for vacancy in vacancies_by_notifications:
-        response = await _make_request_to_negotiations(
-            url=_get_link_to_negotiations(
-                vacancy.pk, user.resume_id, user.covering_letter
-            ),
-            headers=headers,
-        )
-        try:
-            await sync_to_async(_check_response_negotiations)(response)
-            count_valid_requests += 1
-        except AuthError:
-            break
-        except ToContinueError:
-            continue
-        except ToBreakError:
-            break
-    return count_valid_requests
-
-
 async def vacancies(user_id) -> None:
     """Getting vacancies from hh, parse and save to Vacancy model,
     then create Notification model and send to telegram message."""
@@ -235,15 +156,6 @@ async def vacancies(user_id) -> None:
     await _get_vacancies(user)
     info_notification = await sync_to_async(_create_notification)(user)
     await main.send_message_notification(info_notification, user)
-
-
-async def send_negotiations(user_id, id_notification: int) -> None:
-    user = await _get_user(user_id)
-    headers = await sync_to_async(_get_headers)(user.auth_hh.access_token)
-    vacancies_to_negotiations = await sync_to_async(_get_notification)(id_notification)
-    count_valid_negotiations = await _make_negotiations(
-        vacancies_to_negotiations, user, headers
-    )
 
 
 if __name__ == "__main__":
