+++ src/models/models.py
@@ -3,7 +3,7 @@ from phonenumber_field.modelfields import PhoneNumberField
 
 
 class Vacancy(models.Model):
-    user = models.ForeignKey("user.User", on_delete=models.PROTECT)
+    user = models.ManyToManyField("user.User", through="VacanciesInfo")
     company_name = models.CharField(max_length=128)
     name = models.CharField(max_length=127)
     experience = models.CharField(max_length=63)
@@ -12,10 +12,16 @@ class Vacancy(models.Model):
     schedule = models.CharField(max_length=63)
     contact_email = models.EmailField(null=True)
     contacts_phones = PhoneNumberField(null=True)
-    is_request = models.BooleanField(default=False)
 
     def __str__(self):
         return self.name
+
+
+class VacanciesInfo(models.Model):
+    user = models.OneToOneField("user.User", on_delete=models.PROTECT)
+    vacancy = models.ForeignKey(Vacancy, on_delete=models.PROTECT)
+    is_request = models.BooleanField(default=False)
+    date_create = models.DateField(auto_now=True)
 
 
 class Notification(models.Model):
